feat(integrations): refactor metadata for integration specific selection
